all right everything works fine

+++ Ray-Casting/Player.hpp
@@ -24,8 +24,8 @@ public:	//function
 	Player(const int x, const int y);
 	
 	void move(float distanceX, float distanceY){		
-		position.x +=  distanceX * cosf(angle*3.14/180);
-		position.y -=  distanceY * sinf(angle*3.14/180); // on met le '-' car les axes sont inversé ( en bas = y augmente)
+		position.x +=  static_cast<float>(distanceX * cosf(angle*3.14/180));
+		position.y -=  static_cast<float>(distanceY * sinf(angle*3.14/180)); // on met le '-' car les axes sont inversé ( en bas = y augmente)
 	}
